+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Fixed
+- [ V2 ] Fix a sprints calc bug.
+
 ### Added
 - [ V2 ] Dialog confirmation when story state is changed in expanded shape.